Implement K-means clustering and PCA, and use them to compress images.
We implement the K-means clustering algo, including the cluster assignment
stage, the centroid update stage, and the random initialization stage.

After implementing K-means clustering, we apply it to compress images.

We additionally implement PCA. In the first stage, we compute the
covariance matrix of the data. In the second stage, we use singular value
decomposition to compute the eigenvectors (principal components) of the
data.

Finally, we apply PCA to compress (via PCA projection) and reconstruct
(via PCA reconstruction) images.